Revert "Move some classes to sort classes by their feature"
This reverts commit 19b5bd0e40fc32f7e568df9cc7cbceef58cdccd2.
